Update reporting interval right after reporting-day.
Now, the reporting interval would be updated as it is updated in the
daily report. It makes sense because I would usually want to look the
current report to be able to see how I am doing. I cannot change
anything about the last report. If I want to I can always go back in
time within and report and see the last month's report.