Make current entry overlay optional.
It's not that great a feature to highlight the current entry that is
being edited. It's good if only the old entries that I view are using
that overlay. For the time being, I have fixed the overlay to
highlight current line even beyond the last character in the buffer
using 'after-string special property. I have set it to a string that
is display as a huge single space of width just enough to align to the
`(window-text-width)`.

This thing messes up with how `company-mode' completions are displayed
and I just don't like that.