Define hooks pre/post hledger input commit to journal
This makes it easier for me to see the new balances of the accounts
which were affected by my new entries. For someone, trying to do the
same, it is required that they add `hledger-show-new-balances' to
their `hledger-input-post-commit-hook'. Currently the balances are
shown as plain `messages'. I might make it a bit fancier later on.