Change meaning of monthly savings
Monthly savings used to change in assets (assumed to be positive) over
the past one year. I recently saw that this changes more abruptly
compared to monthly income and expenses (I am not sure why exactly
this is happening). This is what I have observed. So, I have decided
that I would consider the difference between monthly income and
monthly expenses to be my monthly savings.