Fix syntax highlighting with other currency types
It seems that previously, when changing the hledger-currency-string,
the regular expression to match amounts was not updated, as it was set
in a variable. This changes the regular expression to match amounts to
a function, which correctly takes into consideration an updated
currency symbol. I believe this may fix #23 and #24.